+++ apps/api/src/modules/archived-notifications/entities/archived-notification.entity.ts
@@ -30,7 +30,7 @@ export class ArchivedNotification {
   @Field()
   providerId: number;
 
-  @Column({ name: 'channel_type', type: 'tinyint', width: 1, default: null })
+  @Column({ name: 'channel_type', type: 'smallint', width: 1, default: null })
   @IsEnum(ChannelType)
   @Field()
   channelType: number;
@@ -42,7 +42,7 @@ export class ArchivedNotification {
 
   @Column({
     name: 'delivery_status',
-    type: 'tinyint',
+    type: 'smallint',
     width: 1,
     default: DeliveryStatus.PENDING,
   })
@@ -73,7 +73,7 @@ export class ArchivedNotification {
   updatedBy: string;
 
   @Column({
-    type: 'tinyint',
+    type: 'smallint',
     width: 1,
     default: Status.ACTIVE,
   })
@@ -85,7 +85,7 @@ export class ArchivedNotification {
   @Field()
   applicationId: number;
 
-  @Column({ name: 'retry_count', type: 'int', default: 0 })
+  @Column({ name: 'retry_count', type: 'integer', default: 0 })
   @Field()
   retryCount: number;
